web-app/backend/requirements.txt
email-validator>=2.2.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
pytest>=8.0.0
requests>=2.31.0
werkzeug>=3.0.0
